% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{collate}
\alias{collate}
\alias{collate.character}
\title{Collate simulation results from multiple scenarios}
\usage{
collate(x, ..., outputs = NULL)

\method{collate}{character}(
  x,
  ...,
  outputs = NULL,
  .path = file.path("cts_results", as.Date(Sys.time()))
)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments (currently ignored).}

\item{outputs}{(optional) A \link{cts_output} or \link{cts_output_list}.}

\item{.path}{The path containing the simulation results. This directory
should contain the file \code{scenarios.csv} as well as subfolders for each
scenario, as generated by \code{\link[=CTS]{CTS()}}.}
}
\value{
A \code{data.frame}.
}
\description{
This function can be used after running multiple scenarios with \code{CTS} to
combine the results for further processing and output generation.
}
