% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{counter-functions}
\alias{counter-functions}
\alias{next_counter}
\alias{set_counter}
\alias{reset_counter}
\alias{list_counters}
\title{Counters}
\usage{
next_counter(counter)

set_counter(counter, value)

reset_counter(counter, .all = FALSE)

list_counters()
}
\arguments{
\item{counter}{A character giving the name of the counter.}

\item{value}{Any value or object.}

\item{.all}{A logical. Should all counters be included?}
}
\value{
\code{next_counter()} returns an integer.
}
\description{
Counters implemented as closures. Each time \code{next_counter()} is called,
it increments the counter.
}
\examples{
next_counter("mycounter")
next_counter("mycounter")
next_counter("mycounter")
list_counters()
set_counter("mycounter", 10)
reset_counter("mycounter")
}
