% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{CTS}
\alias{CTS}
\title{Perform a clinical trial simulation}
\usage{
CTS(x, ...)
}
\arguments{
\item{x}{An object (e.g. a \link{scenario} or \link{scenario_list}).}

\item{...}{Further arguments, ignored.}
}
\description{
Perform a clinical trial simulation
}
