% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{cts_module_}
\alias{cts_module_}
\alias{parameters_}
\alias{population_}
\alias{subject_model_}
\alias{treatment_}
\alias{pkpd_model_}
\alias{simulation_}
\alias{endpoints_}
\alias{summarization_}
\title{CTS module (pipe interface)}
\usage{
parameters_(scenario, ...)

population_(scenario, ...)

subject_model_(scenario, ...)

treatment_(scenario, ...)

pkpd_model_(scenario, ...)

simulation_(scenario, ...)

endpoints_(scenario, ...)

summarization_(scenario, ...)
}
\description{
CTS module (pipe interface)
}
