% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{cts_module_grid}
\alias{cts_module_grid}
\alias{parameters_grid}
\alias{population_grid}
\alias{subject_model_grid}
\alias{treatment_grid}
\alias{pkpd_model_grid}
\alias{simulation_grid}
\alias{endpoints_grid}
\alias{summarization_grid}
\title{CTS module grid}
\usage{
parameters_grid(...)

population_grid(...)

subject_model_grid(...)

treatment_grid(...)

pkpd_model_grid(...)

simulation_grid(...)

endpoints_grid(...)

summarization_grid(...)
}
\arguments{
\item{...}{Any objects (typically vectors or lists).}
}
\value{
A \code{list} of \link{cts_module} objects.
}
\description{
CTS module grid
}
\examples{
# This will create 5 scenarios, each with a different value for the parameter x:
scenario_grid(scenario(name="foo"), parameters=parameters_grid(x=1:5))
}
