% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{cts_output}
\alias{cts_output}
\alias{csv_output}
\alias{graphical_output}
\title{CTS outputs}
\usage{
csv_output(x, ..., name)

graphical_output(x, ..., name, width, height)
}
\arguments{
\item{x}{An expression.}

\item{...}{Additional arguments (currently ignored).}

\item{name}{A name. It is used to construct the filename for the saved output.}

\item{width}{For graphical outputs, the width in inches.}

\item{height}{For graphical outputs, the height in inches.}
}
\value{
An object that inherits from the class \code{cts_output}.
}
\description{
Modules are the components that make up a CTS scenario. The power of the
framework comes from the ability to mix and match modules to easily create a
multitude of scenarios.
}
\details{
Each module has the same structure: it consists of a named list of
expressions that are run in sequence when the scenario is executed.
}
