% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{get_scenario_table}
\alias{get_scenario_table}
\title{Get the scenario table for a list of scenarios}
\usage{
get_scenario_table(scenario_list)
}
\arguments{
\item{scenario_list}{The \link{scenario_list} for which to get the table.}
}
\value{
A \code{data.frame}.
}
\description{
Get the scenario table for a list of scenarios
}
