% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{output_list}
\alias{output_list}
\alias{cts_output_list}
\title{CTS output list}
\usage{
output_list(...)
}
\arguments{
\item{...}{Any number of \link{cts_output} objects.}
}
\value{
An object that inherits from the class \code{cts_output_list}.
}
\description{
An object that stores a list of \link{cts_output} objects. Used when multiple
outputs are desired for a single module.
}
