% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{parameter_uncertainty}
\alias{parameter_uncertainty}
\title{Add parameter uncertainty to a scenario}
\usage{
parameter_uncertainty(scenario, ...)
}
\arguments{
\item{scenario}{The \link{scenario} to add parameter uncertainty to.}
}
\description{
Add parameter uncertainty to a scenario
}
