% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{qi}
\alias{qi}
\title{Quantile interval}
\usage{
qi(x, level = 0.95, type = 7)
}
\arguments{
\item{x}{A numeric vector.}

\item{level}{The level of the interval, typically 0.95, 0.9 or 0.5 (which
corresponds to the interquartile range).  Determines the quantiles that
correspond to the endpoints of the interval.  Can also be a vector to
compute multiple intervals at once.}

\item{type}{The quantile algorithm (see \code{\link[stats:quantile]{stats::quantile()}}).}
}
\value{
A named list of numbers, giving the lower and upper limits of the
interval, as well as the median.
}
\description{
A level \eqn{\alpha} quantile interval of \code{x} is the interval whose endpoints are the
upper and lower \eqn{(1-\alpha)/2} quantiles of \code{x}. In other words, it covers a
\eqn{(1-\alpha)} fraction of the data, with an equal number of points outside the
interval in each tail.
}
\examples{
x <- 1:1000
qi(x, 0.9)
qi(x, c(0.95, 0.9, 0.5))
}
