% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{replication}
\alias{replication}
\title{Scenario replication}
\usage{
replication(scenario, nreplicates, seed = NULL, collate = NULL)
}
\arguments{
\item{scenario}{The \link{scenario} to replicate.}

\item{seed}{The initial RNG seed to use.}

\item{collate}{Collate the replication results.}

\item{nreplicatesd}{The number of times to replicate the scenario.}
}
\description{
Scenario replication
}
