% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{scenario}
\alias{scenario}
\title{Define a CTS scenario}
\usage{
scenario(
  parameters = NULL,
  population = NULL,
  subject_model = NULL,
  treatment = NULL,
  pkpd_model = NULL,
  simulation = NULL,
  endpoints = NULL,
  summarization = NULL,
  ...,
  name = NULL,
  desc = NULL,
  seed = make_seed(),
  .silent = FALSE
)
}
\arguments{
\item{parameters}{\link{parameters} module (optional).}

\item{population}{\link{population} module (optional).}

\item{subject_model}{\link{subject_model} module (optional).}

\item{treatment}{\link{treatment} module (optional).}

\item{pkpd_model}{\link{pkpd_model} module (optional).}

\item{simulation}{\link{simulation} module (optional).}

\item{endpoints}{\link{endpoints} module (optional).}

\item{summarization}{\link{summarization} module (optional).}

\item{...}{Additional arguments, ignored.}

\item{name}{Optional scenario name.}

\item{desc}{Optional scenario description.}

\item{seed}{RNG seed, for reproducibility.}

\item{.silent}{Set to \code{TRUE} to supress informative messages.}
}
\description{
The scenario is the central concept of the CTS framework. A scenario is
defined by one or more modules. It is the ability to mix and match modules
to create a multitute of scenarios that gives the framework its power.
}
