% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{scenario_grid}
\alias{scenario_grid}
\title{Create a scenario grid}
\usage{
scenario_grid(
  base_scenario,
  parameters = NULL,
  population = NULL,
  subject_model = NULL,
  treatment = NULL,
  pkpd_model = NULL,
  simulation = NULL,
  endpoints = NULL,
  summarization = NULL,
  ...
)
}
\arguments{
\item{base_scenario}{A \link{scenario} which serves as the base scenario from which to derive.}

\item{parameters}{\link{parameters} module (optional).}

\item{population}{\link{population} module (optional).}

\item{subject_model}{\link{subject_model} module (optional).}

\item{treatment}{\link{treatment} module (optional).}

\item{pkpd_model}{\link{pkpd_model} module (optional).}

\item{simulation}{\link{simulation} module (optional).}

\item{endpoints}{\link{endpoints} module (optional).}

\item{summarization}{\link{summarization} module (optional).}

\item{...}{Further argument passed to \code{\link[=scenario]{scenario()}}.}
}
\value{
A \code{cts_scenario_list} object.
}
\description{
A scenario grid is like a \link{scenario_list} except that all combinations of
multiple different factors are generated.
}
\examples{
# A silly example that does nothing. For realistic examples, see the vignette.

# This will create 12 scenarios in total, representing all
# combinations of x, y and treatment.
CTS(scenario_grid(
    scenario(name="base", simulation=simulation({cat(x, y, treatment, "\n")})),
    parameters=parameters_grid(x=1:2, y=3:5),
    treatment=list(
        treatment("A"),
        treatment("B")
    )
))
}
