% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{update.cts_scenario}
\alias{update.cts_scenario}
\title{Update an existing scenario}
\usage{
\method{update}{cts_scenario}(object, ..., .silent = TRUE)
}
\arguments{
\item{object}{The \link{scenario} to be updated.}

\item{...}{Further arguments passed to \code{\link[=derive_scenario]{derive_scenario()}}.}

\item{.silent}{Set to \code{TRUE} to supress informative messages.}
}
\description{
Update an existing scenario
}
