% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{collate}
\alias{collate}
\title{Collate simulation results from multiple scenarios}
\usage{
collate(x, ..., outputs = NULL)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments (ignored).}

\item{outputs}{(optional) A \link{cts_output} or \link{cts_output_list}.}
}
\value{
A \code{data.frame}
}
\description{
Collate simulation results from multiple scenarios
}
