% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{cts_module}
\alias{cts_module}
\alias{parameters}
\alias{population}
\alias{subject_model}
\alias{treatment}
\alias{pkpd_model}
\alias{simulation}
\alias{endpoints}
\alias{summarization}
\title{CTS modules}
\usage{
parameters(..., outputs = NULL, seed = NULL)

population(..., outputs = NULL, seed = NULL)

subject_model(..., outputs = NULL, seed = NULL)

treatment(..., outputs = NULL, seed = NULL)

pkpd_model(..., outputs = NULL, seed = NULL)

simulation(..., outputs = NULL, seed = NULL)

endpoints(..., outputs = NULL, seed = NULL)

summarization(..., outputs = NULL, seed = NULL)
}
\arguments{
\item{...}{A named list of expressions.}

\item{outputs}{(optional) A \link{cts_output} or \link{cts_output_list}.}

\item{seed}{Modules can be given a RNG seed, for reproducibility.}
}
\value{
A module object.
}
\description{
Modules are the components that make up a CTS scenario. The power of the
framework comes from the ability to mix and match modules to easily create a
multitute of scenarios.
}
\details{
Each module has the same structure: it consists of a named list of
expressions that are run in sequence when the scenario is executed.
}
