% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{replication}
\alias{replication}
\title{Scenario replication}
\usage{
replication(scenario, nreplicates, collate = NULL)
}
\arguments{
\item{scenario}{The \link{scenario} to replicate.}

\item{nreplicates}{The number of times to replicate the scenario.}

\item{collate}{A \link{collate} object for combining results across replicates.}
}
\value{
A \link{scenario}.
}
\description{
Scenario replication
}
\examples{
# A simple power calculation example
power_scenario <- scenario(
    simulation = simulation(
        t.test(rnorm(10, mean=1), rnorm(10, mean=2))$p.value
    )
) |> replication(nreplicates=1000,
    collate = collate(
        power = mean(sapply(replicates, getElement, name="simulation") < 0.05
    )
))

x <- CTS(power_scenario)
x$power
}
