% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{scenario}
\alias{scenario}
\title{Define a CTS scenario}
\usage{
scenario(
  parameters = NULL,
  population = NULL,
  subject_model = NULL,
  treatment = NULL,
  pkpd_model = NULL,
  simulation = NULL,
  endpoints = NULL,
  summarization = NULL,
  ...,
  name = NULL,
  seed = make_seed(),
  .silent = FALSE
)
}
\arguments{
\item{parameters}{\link{parameters} module (optional).}

\item{population}{\link{population} module (optional).}

\item{subject_model}{\link{subject_model} module (optional).}

\item{treatment}{\link{treatment} module (optional).}

\item{pkpd_model}{\link{pkpd_model} module (optional).}

\item{simulation}{\link{simulation} module (optional).}

\item{endpoints}{\link{endpoints} module (optional).}

\item{summarization}{\link{summarization} module (optional).}

\item{...}{Additional arguments (currently ignored).}

\item{name}{Optional scenario name. If omitted, a default name will be
generated such as \code{ScenarioX} where \code{X} is a number that is incremented for
each new scenario created.}

\item{seed}{RNG seed, for reproducibility.}

\item{.silent}{Set to \code{TRUE} to suppress informative messages.}
}
\value{
A \code{cts_scenario} object.
}
\description{
The scenario is the central concept of the CTS framework. A scenario is
defined by one or more modules. It is the ability to mix and match modules
to create a multitude of scenarios that gives the framework its power.
}
\examples{
# A silly example that does nothing. For realistic examples, see the vignette.
CTS(scenario(name="foo",
    parameters    = parameters(),
    population    = population(),
    subject_model = subject_model(),
    treatment     = treatment(),
    pkpd_model    = pkpd_model(),
    simulation    = simulation(),
    endpoints     = endpoints(),
    summarization = summarization()
))

# There is also a syntax that uses pipes to build up the scenario
# (note the mandatory underscore '_' at the end of each module name):
scenario(name="foo") |>
    parameters_() |>
    population_() |>
    subject_model_() |>
    treatment_() |>
    pkpd_model_() |>
    simulation_() |>
    endpoints_() |>
    summarization_() |>
    CTS()

}
