% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsframework.R
\name{scenario_list}
\alias{scenario_list}
\title{Create a scenario list}
\usage{
scenario_list(..., collate = NULL)
}
\arguments{
\item{...}{Any number of \link{scenario}s.}
}
\description{
Multiples scenarios are combined into a list that can be run as a unit.
}
